backend/requirements.txt
asgiref==3.7.2
astroid==2.15.6
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
class-registry==2.1.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
Django==3.2.3
django-appconf==1.0.5
django-colorfield==0.9.0
django-cors-headers==4.1.0
django-extra-fields==3.0.2
django-filter==23.2
django-measurement==3.2.4
django-templated-mail==1.1.1
DjangoHexadecimalField==0.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docformatter==1.7.5
filters==1.3.2
flake8==6.0.0
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
measurement==3.2.2
mpmath==1.3.0
oauthlib==3.2.2
pep8radius==0.9.2
Pillow==9.5.0
platformdirs==3.8.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pylint==2.17.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
sympy==1.12
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
untokenize==0.1.1
uritemplate==4.1.1
urllib3==2.0.3
webcolors==1.13
wrapt==1.15.0
yapf==0.40.1
zipp==3.16.0
